Refactored 'value' to 'latestValue' in Updatable
Added LocalDataSync. Unusable at the moment.

RemoteLocator is now Updatable returning a List of Devices.

More fully Tekdaqc utilities for connecting and controlling.

Added TekdaqcLocator wrapper as a generic wrapper for finding Tekdaqcs

Added Unused LocalDataTest

Added TekdaqcTest to test the basic functionality of the Tekdaqc Utilities

TriggerTest now more verbose.

+++ src/main/kotlin/com/tenkiv/daqc/hardware/definitions/channel/DigitalOutput.kt
@@ -13,6 +13,6 @@ abstract class DigitalOutput: Output<DaqcValue.Boolean>,
 
     abstract val canPulseWidthModulate: Boolean
 
-    abstract fun pulseWidthModulate(dutyCycle: Float)
+    abstract fun pulseWidthModulate(dutyCycle: Int)
 
 }
